+++ fase_2/ecommerce_recommender/src/recommender/training/__init__.py
@@ -1,6 +1,9 @@
 """Public training API."""
 
+from .checkpoint import load_checkpoint, save_checkpoint
 from .early_stopping import EarlyStopping
+from .evaluator import compute_ranking_metrics
+from .experiment import ExperimentConfig, ExperimentResult, train_one_experiment
 from .metrics import hit_rate_at_k, ndcg_at_k
 from .trainer import EpochResult, Trainer
 
@@ -8,6 +11,12 @@ __all__ = [
     "EarlyStopping",
     "EpochResult",
     "Trainer",
+    "ExperimentConfig",
+    "ExperimentResult",
+    "train_one_experiment",
+    "compute_ranking_metrics",
     "hit_rate_at_k",
     "ndcg_at_k",
+    "load_checkpoint",
+    "save_checkpoint",
 ]